#14 - Extract some code into separate folders for easier T4 solution integration.
