+++ src/IX.Retry/StandardPolicies/TimeBasedRetryPolicy.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace IX.Retry
+namespace IX.Retry.StandardPolicies
 {
     /// <summary>
     /// A time-based, non-linear retry policy.
@@ -12,6 +12,7 @@ namespace IX.Retry
         private readonly TimeSpan retryDuration;
         private readonly IEnumerable<Type> retryForExceptionTypes;
         private DateTime start;
+        private readonly BackOffPolicy backOffPolicy;
 
         /// <summary>
         /// Instantiates a new instance of the <see cref="TimeBasedRetryPolicy"/> class.
@@ -22,9 +23,24 @@ namespace IX.Retry
         /// <para>The exception types are not checked, meaning that if a collection of types that are not exceptions is given as parameter, then the retry mechanism will do nothing.</para>
         /// </remarks>
         public TimeBasedRetryPolicy(TimeSpan retryDuration, IEnumerable<Type> retryForExceptionTypes)
+            : this(retryDuration, StandardBackoffPolicies.RandomExponential, retryForExceptionTypes)
+        {
+        }
+
+        /// <summary>
+        /// Instantiates a new instance of the <see cref="TimeBasedRetryPolicy"/> class.
+        /// </summary>
+        /// <param name="retryDuration">The retry duration for this time-based policy.</param>
+        /// <param name="backOffPolicy">The back-off policy to use.</param>
+        /// <param name="retryForExceptionTypes">A collection of exception types for which to retry.</param>
+        /// <remarks>
+        /// <para>The exception types are not checked, meaning that if a collection of types that are not exceptions is given as parameter, then the retry mechanism will do nothing.</para>
+        /// </remarks>
+        public TimeBasedRetryPolicy(TimeSpan retryDuration, BackOffPolicy backOffPolicy, IEnumerable<Type> retryForExceptionTypes)
         {
             this.retryDuration = retryDuration;
             this.retryForExceptionTypes = retryForExceptionTypes;
+            this.backOffPolicy = backOffPolicy;
         }
 
         /// <summary>
@@ -59,7 +75,7 @@ namespace IX.Retry
                 return DontRetry(out retryInterval);
             }
 
-            retryInterval = StandardBackoffPolicies.RandomExponential(retryCount, Policy.MinBackoff, Policy.MaxBackoff, Policy.BackoffIncrement);
+            retryInterval = backOffPolicy(retryCount, Policy.MinBackoff, Policy.MaxBackoff, Policy.BackoffIncrement);
             return true;
         }
     }
